Move implementation details to `internal` mod
